backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.3
python-dotenv>=1.0.1
python-multipart>=0.0.6
passlib>=1.7.4
bcrypt>=4.1.2
python-jose>=3.3.0
mysql-connector-python>=8.2.0
pymysql>=1.1.0
cryptography
email-validator
reportlab
cloudinary
openpyxl
